Move Reflection material to materials index
