+++ js/components/3D/Materials/index.jsx
@@ -1,8 +1,13 @@
 import * as THREE from "three";
 import { useControls, folder } from "leva";
 import { useRef } from "react";
-import { ItemShaderMaterial, TriptychShaderMaterial } from "../Shaders";
+import {
+    ItemShaderMaterial,
+    TriptychShaderMaterial,
+    ReflectionShaderMaterial,
+} from "../Shaders";
 import * as colors from "@/js/core/colors";
+import { useFrame } from "@react-three/fiber";
 
 export const TriptychMaterial = ({ boundingBox, opacity = 1 }) => {
     const shaderRef = useRef();
@@ -97,6 +102,155 @@ export const TriptychMaterial = ({ boundingBox, opacity = 1 }) => {
             uGradientEnabled={uGradientEnabled}
             ref={shaderRef}
             key={TriptychShaderMaterial.key}
+            opacity={opacity}
+            transparent
+        />
+    );
+};
+
+export const ReflectionMaterial = ({ boundingBox, opacity = 1 }) => {
+    const shaderRef = useRef();
+    const time = useRef(1.0);
+    const {
+        uPerlinEnabled,
+        uPerlinResolution,
+        uPerlinYScale,
+        uPerlinSpeed,
+        uPerlinLightenFactor,
+        uNoiseScale,
+        uNoiseContrast,
+        uNoiseScalarDistanceFactor,
+        uGradientStop,
+        uMatcapEnabled,
+        uNoiseEnabled,
+        uGradientEnabled,
+        uClampColorEnabled,
+        uClampColorMax,
+        uClampColorMin,
+    } = useControls({
+        "Reflection Shader": folder({
+            Perlin: folder({
+                uPerlinEnabled: true,
+                uPerlinResolution: {
+                    value: 8,
+                    min: 1,
+                    max: 50,
+                    step: 1,
+                },
+                uPerlinLightenFactor: {
+                    value: 0.3,
+                    min: 0,
+                    max: 1,
+                    step: 0.1,
+                },
+                uPerlinYScale: {
+                    value: 6.0,
+                    min: 1.0,
+                    max: 10.0,
+                    step: 1.0,
+                },
+                uPerlinSpeed: {
+                    value: 0.5,
+                    min: 0.1,
+                    max: 10,
+                    step: 0.1,
+                },
+            }),
+            Matcap: folder({
+                uMatcapEnabled: true,
+            }),
+            Noise: folder({
+                uNoiseEnabled: true,
+                uNoiseScale: {
+                    value: 850,
+                    min: 10,
+                    max: 2000,
+                    step: 10,
+                },
+                uNoiseScalarDistanceFactor: {
+                    value: 0.9,
+                    min: 0,
+                    max: 10,
+                    step: 0.1,
+                },
+                uNoiseContrast: {
+                    value: 1.0,
+                    min: 0,
+                    max: 1,
+                    step: 0.1,
+                },
+            }),
+            Gradient: folder({
+                uGradientEnabled: true,
+                uGradientStop: {
+                    value: 0.21,
+                    min: 0.0,
+                    max: 1.0,
+                    step: 0.01,
+                },
+            }),
+            ColorClamp: folder({
+                uClampColorEnabled: true,
+                uClampColorMin: {
+                    value: {
+                        r: 52,
+                        g: 52,
+                        b: 52,
+                    },
+                    onChange: (v) => {
+                        const color = new THREE.Vector3(
+                            v.r,
+                            v.g,
+                            v.b,
+                        ).divideScalar(255);
+                        shaderRef.current.uniforms.uClampColorMin.value = color;
+                    },
+                },
+                uClampColorMax: {
+                    value: {
+                        r: 255,
+                        g: 255,
+                        b: 255,
+                    },
+                    onChange: (v) => {
+                        const color = new THREE.Vector3(
+                            v.r,
+                            v.g,
+                            v.b,
+                        ).divideScalar(255);
+                        shaderRef.current.uniforms.uClampColorMax.value = color;
+                    },
+                },
+            }),
+        }),
+    });
+
+    useFrame(({ clock }) => {
+        time.current = clock.getElapsedTime();
+        shaderRef.current.uniforms.time.value = time.current;
+    });
+
+    return (
+        <reflectionShaderMaterial
+            uPerlinEnabled={uPerlinEnabled}
+            uPerlinResolution={uPerlinResolution}
+            uPerlinSpeed={uPerlinSpeed}
+            uPerlinYScale={uPerlinYScale}
+            uPerlinLightenFactor={uPerlinLightenFactor}
+            uNoiseEnabled={uNoiseEnabled}
+            uNoiseScale={uNoiseScale}
+            uNoiseScalarDistanceFactor={uNoiseScalarDistanceFactor}
+            uNoiseContrast={uNoiseContrast}
+            uBoundingBoxMin={boundingBox.min}
+            uBoundingBoxMax={boundingBox.max}
+            uGradientStop={uGradientStop}
+            uMatcapEnabled={uMatcapEnabled}
+            uGradientEnabled={uGradientEnabled}
+            uClampColorEnabled={uClampColorEnabled}
+            uClampColorMax={uClampColorMax}
+            uClampColorMin={uClampColorMin}
+            ref={shaderRef}
+            key={ReflectionShaderMaterial.key}
             opacity={opacity}
             transparent
         />
